pyproject.toml
[tool.poetry]
name = "agentic-resume-matcher"
version = "0.1.0"
description = "Production-ready resume tailoring system using Strands Agents SDK"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
boto3 = "^1.34.0"
PyPDF2 = "^3.0.0"
python-docx = "^1.0.0"
reportlab = "^4.0.0"
click = "^8.1.0"
rich = "^13.0.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.12.0"
black = "^23.0.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
isort = "^5.12.0"

[tool.poetry.group.api]
optional = true

[tool.poetry.group.api.dependencies]
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
